refactor(proof): move Proof struct to data crate
To be used by both the PVM and NDS
