+++ pvm/src/state_backend/proof_backend.rs
@@ -6,4 +6,3 @@
 //! Proof-related functionality
 
 pub mod merkle;
-pub mod proof;
