+++ pvm/tests/test_proofs.rs
@@ -13,8 +13,6 @@ use std::time::Instant;
 use octez_riscv::machine_state::memory::M64M;
 use octez_riscv::machine_state::memory::MemoryConfig;
 use octez_riscv::pvm::hooks::NoHooks;
-use octez_riscv::state_backend::proof_backend::proof::Proof;
-use octez_riscv::state_backend::proof_backend::proof::serialise_proof;
 use octez_riscv::state_backend::verify_backend::ProofVerificationFailure;
 use octez_riscv::stepper::Stepper;
 use octez_riscv::stepper::StepperStatus;
@@ -22,6 +20,8 @@ use octez_riscv::stepper::pvm::PvmStepper;
 use octez_riscv_data::foldable::Foldable;
 use octez_riscv_data::hash;
 use octez_riscv_data::hash::HashFold;
+use octez_riscv_data::merkle_proof::proof::Proof;
+use octez_riscv_data::merkle_proof::proof::serialise_proof;
 use octez_riscv_data::merkle_proof::proof_tree::MerkleProofFold;
 use octez_riscv_data::mode::Normal;
 use octez_riscv_data::mode::Prove;
@@ -280,8 +280,8 @@ fn basic_invalid_proofs_are_rejected<MC: MemoryConfig>(
 }
 
 mod proof_helpers {
-    use octez_riscv::state_backend::proof_backend::proof::Proof;
     use octez_riscv_data::hash::Hash;
+    use octez_riscv_data::merkle_proof::proof::Proof;
     use octez_riscv_data::merkle_proof::proof_tree::MerkleProofLeaf;
     use octez_riscv_data::tree::Tree;
